fixed authentication

+++ src/main/java/com/vivekgude/leastcount/service/GameService.java
@@ -5,7 +5,7 @@ import com.vivekgude.leastcount.model.dto.GameDTO;
 import java.util.Optional;
 
 public interface GameService {
-    GameDTO createGame(int userId, String userName);
+    GameDTO createGame(long userId, String userName);
 
-    Optional<GameDTO> joinGame(int userId, String userName, String gameId);
+    Optional<GameDTO> joinGame(long userId, String userName, String gameId);
 }
